refactor cursor modules
