+++ CHANGES
@@ -1,18 +1,18 @@
 main
 ====
-Cursors (class hierarchy revised for findAndRerank, plus other changes):
+Cursors (class hierarchy revised to accommodate `find_and_rerank`, plus other changes):
     - renamed the 'FindCursorState' enum to `CursorState`
     - renamed the abstract ur-class 'FindCursor' => `AbstractCursor`
     - `clone` method does not strip the mapping anymore, rather retains it
     - removed the `alive` sugar property (use `cursors.state != CursorState.CLOSED`)
 Support for findAndRerank in collections:
-    - new classes `CollectionLexicalOptions`, `CollectionRerankingOptions`, `RerankingServiceOptions`
+    - new classes `CollectionLexicalOptions`, `CollectionRerankingOptions`, `RerankingServiceOptions` for create_collection
     - new `lexical` and `reranking` entries in CollectionDefinition (+builder interface management)
     - `find_and_rerank` method for collections
     - cursor classes `[Async]CollectionFindAndRerankCursor` added
     - findAndRerank cursors return the new `RerankedResult` construct by default (modulo custom mappings)
 Maps for tables expressed as list of pairs (association lists):
-    - support for automatic handling of dicts and DataAPIMaps in the proper table payload portions
+    - support for automatic handling of DataAPIMaps (+possibly dicts) in the proper table payload portions
     - introduced serdes option `encode_maps_as_lists_in_tables` (default to "NEVER") to control this
 Exceptions, major rework of `[Table|Collection]InsertManyException`, `CollectionUpdateManyException` and `CollectionDeleteManyException`
     - All astrapy exceptions derive directly from `Exception` (and not 'ValueError` anymore)
@@ -52,6 +52,7 @@ Replaced the ValueErrors not directly coming from function calls/constructors wi
 Collection and Table `insert_many` methods employ returnDocumentResponses under the hood.
 maintenance: switch to DSE6.9 for local non-Astra testing.
 
+
 v 2.0.0rc1
 ==========
 Support for TIMEUUID and COUNTER columns:
@@ -63,6 +64,7 @@ added optional `api_support` member to all column descriptor (as the Data API re
 restore support for Python 3.8, 3.9
 maintenance: full restructuring of tests and CI (tables+collections on same footing+other)
 maintenance: adopt `blockbuster` in async tests to detect (and bust) any blocking call
+
 
 v 2.0.0-preview
 ===============
@@ -221,6 +223,7 @@ Internal restructuring/maintenance things:
     - major restructuring of the codebase in directories (some internal-only imports changed; reduced the scope of `test_imports`)
     - removal of unused imports from toplevel `__init__.py` (ids, constants, cursors)
     - simplified timeout management classes and representations
+
 
 v. 1.5.2
 ========
